[GStreamer] Refactor media player to use MediaTime consistently
Backport of https://bugs.webkit.org/show_bug.cgi?id=174817 and some fixes
on top of it to make it build on WPEWebKit master.

Original author: Charlie Turner <cturner@igalia.com>

+++ Source/WebCore/platform/graphics/gstreamer/mse/PlaybackPipeline.cpp
@@ -494,8 +494,8 @@ void PlaybackPipeline::enqueueSample(RefPtr<MediaSample> mediaSample)
     GST_TRACE("enqueing sample trackId=%s PTS=%f presentationSize=%.0fx%.0f at %" GST_TIME_FORMAT " duration: %" GST_TIME_FORMAT,
         trackId.string().utf8().data(), mediaSample->presentationTime().toFloat(),
         mediaSample->presentationSize().width(), mediaSample->presentationSize().height(),
-        GST_TIME_ARGS(WebCore::toGstClockTime(mediaSample->presentationTime().toDouble())),
-        GST_TIME_ARGS(WebCore::toGstClockTime(mediaSample->duration().toDouble())));
+        GST_TIME_ARGS(WebCore::toGstClockTime(mediaSample->presentationTime())),
+        GST_TIME_ARGS(WebCore::toGstClockTime(mediaSample->duration())));
 
     Stream* stream = getStreamByTrackId(m_webKitMediaSrc.get(), trackId);
